Patch submitted by Christian Packmann (thanks!)
- BUGFIX: deadkeys work. The use of ToAscii() in ConvertToAscii() seems
to modify Windows' internal keyboard state, preventing the processing of
dead keys. I replaced ToAscii() with MapVirtualKey(), this seems to give
identical results for purposes of ASCII translation, but deadkeys and
Alt-Numpad Codepage/Unicode entry work now (on my German keyboard, anyway).
This should close bug
http://sourceforge.net/tracker/index.php?func=detail&aid=1624870&group_id=65529&atid=511317

- BUGFIX: building single-file .exes with py2exe works now. I've adapted
the patch
http://sourceforge.net/tracker/index.php?func=detail&aid=1532828&group_id=65529&atid=511319
for this, but changed the behavior. The patch would *only* work in .exes
as it would only try to get the module handle for the currently running .exe, which has to fail if a script is executed by python.exe.
The code now tries to find the module in the running .exe, if that fails
it tries to load it externally.
I've tested this with my own pyHook project, and it works well as .py or
.exe. This may also solve the problems with PyInstaller, but I haven't tested this.

- internal documentation changes:
* I've added a comment to the docs of HookManager.SubscribeKeyChar()
to indicate that it doesn't work in the current version - at least I've
never seen any intercepted WM_CHAR even if using SubscribeKeyAll().
I think that intercepting WM_*CHAR messages will require an additional
hook (probably WH_MSGFILTER), the current keyboard hook only intercepts
the low-level UP/DOWN keystrokes, not the actual characters generated by
the keystrokes.

- changes to README.txt
* Added descriptions for the bugfixes to "Change log"
* New section "Known bugs"
* New section "Limitations": mention that pyHook won't work on Win9x as
the used low-level hooks are only supported by WinNT 4.0 SP3 and higher.
* Added link to Sourceforge page for reporting bugs/feature requests.